// src/modules/blog/blog.enum.ts
export enum BlogLanguages {
	Farsi = "fa",
	English = "en",
}

export enum BlogStatus {
	Inprogress = "Inprogress",
	Accepted = "Accepted",
	Rejected = "Rejected",
	TopBlog = "TopBlog",
}

export enum BlogLevel {
	Potential = "Potential",
	Augmented = "Augmented",
	Expected = "Expected",
	Generic = "Generic",
	Core = "Core",
}